# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "coco-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/sys/configs": {
      "get": {
        "tags": [
          "System"
        ],
        "operationId": "list_configs",
        "parameters": [
          {
            "name": "org_id",
            "in": "query",
            "description": "Organization identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "description": "Project identifier",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Indexing configs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListConfigsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "System"
        ],
        "operationId": "upsert_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Indexing config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexingConfigResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sys/configs/activate": {
      "post": {
        "tags": [
          "System"
        ],
        "operationId": "activate_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Activated config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivateConfigResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sys/prune": {
      "post": {
        "tags": [
          "System"
        ],
        "operationId": "prune_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PruneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Prune completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PruneResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sys/register": {
      "post": {
        "tags": [
          "System"
        ],
        "operationId": "register_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Project registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterProjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActivateConfigRequest": {
        "type": "object",
        "required": [
          "org_id",
          "project_id",
          "config_id"
        ],
        "properties": {
          "config_id": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          }
        }
      },
      "ActivateConfigResponse": {
        "type": "object",
        "required": [
          "active_config_id"
        ],
        "properties": {
          "active_config_id": {
            "type": "string"
          }
        }
      },
...
      "CocoErrorKind": {
        "type": "string",
        "description": "High-level error categories for CoCo.\n\nAdding new variants is a breaking change for the public API.",
        "enum": [
          "system",
          "user",
          "network",
          "storage",
          "compute"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Serializable error response used by API layers.",
        "required": [
          "kind",
          "message"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/CocoErrorKind"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "IndexingConfigResponse": {
        "type": "object",
        "required": [
          "config"
        ],
        "properties": {
          "config": {
            "$ref": "#/components/schemas/IndexingConfig"
          }
        }
      },
...
      "ListConfigsResponse": {
        "type": "object",
        "required": [
          "configs"
        ],
        "properties": {
          "active_config_id": {
            "type": "string",
            "nullable": true
          },
          "configs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndexingConfig"
            }
          }
        }
      },
...
      "PruneRequest": {
        "type": "object",
        "required": [
          "org_id",
          "project_id"
        ],
        "properties": {
          "keep": {
            "type": "integer",
            "nullable": true,
            "minimum": 0
          },
          "org_id": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          }
        }
      },
      "PruneResponse": {
        "type": "object",
        "required": [
          "removed"
        ],
        "properties": {
          "removed": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "RegisterProjectRequest": {
        "type": "object",
        "required": [
          "org_id",
          "name",
          "source_ref"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "org_name": {
            "type": "string",
            "nullable": true
          },
          "platform": {
            "type": "string",
            "nullable": true
          },
          "project_id": {
            "type": "string",
            "nullable": true
          },
          "source_ref": {
            "type": "string"
          }
        }
      },
      "RegisterProjectResponse": {
        "type": "object",
        "required": [
          "project_id",
          "org_id",
          "name",
          "active_config_id"
        ],
        "properties": {
          "active_config_id": {
            "type": "string"
          },
          "active_version_id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          }
        }
      },
...
      "UpsertConfigRequest": {
        "type": "object",
        "required": [
          "org_id",
          "config"
        ],
        "properties": {
          "config": {
            "$ref": "#/components/schemas/IndexingConfig"
          },
          "org_id": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}